\name{bias}
\alias{bias}
\docType{data}
\title{Prediction Errors ("Biases") of Surface Temperature Forecasts}
\description{
Prediction errors of 48-hour ahead MM5 forecasts of surface
temperature measured at 96 different locations in the US Pacific
Northwest on January 3, 2000. The prediction error, or "bias", is
the difference between the forecasted and observed surface
temperature. (MM5 is the fifth-generation Pennsylvania State
University -- National Center for Atmospheric Research Mesoscale
Model.)
}
\usage{bias}
\source{
Data have been kindly provided by the research group of Professor
Clifford Mass in the Department of Atmospheric Sciences at the
University of Washington. Detailed information about the Pacific
Northwest prediction effort and the associated data archive can be
found online at \url{www.atmos.washington.edu/mm5rt/info.html} and
\url{www.atmos.washington.edu/marka/pnw.html}, respectively.
}
\keyword{datasets}
