\name{blackhire}
\alias{blackhire}
\docType{data}
\title{Hiring data for eight professions and two races}
\description{
Number of black and white candidates (hired or rejected) for eight professions.}
\usage{blackhire}
\format{
  An array with 2 rows by 2 columns by 8 levels.
  }

\references{
Gastwirth, J. L.(1984) \emph{Statistical Methods for Analyzing Claims of Employment Discrimination}, 
Industrial and Labor Relations Review, Vol. 38, No. 1. (October 1984), pp. 75-86.\cr
}

\keyword{datasets}
