\name{data1963}
\alias{data1963}
\docType{data}
\title{Population data of ratio of number of senators and 
   representatives to population size in 1963}
\description{
The dataset of ratio of number of senators and representatives 
to population size in 13 districts in the United States in 1963 
(Gastwirth, 1972).
}
\usage{data1963}
\format{
  A data frame with 13 observations on the following 3 variables.
  \describe{
    \item{\code{pop1963}}{population data in 1963}
    \item{\code{sen1963}}{Number of senators in each district in 1963}
    \item{\code{rep1963}}{Number of representatives in each district in 1963}
  }
}
\source{
Gastwirth, J. L.(1972) \emph{The Estimation of the Lorenz Curve and Gini Index}, The Review of Economics and Statistics, Vol. 54, No. 3. (August 1972), pp. 306-316.\cr
}
\references{
Gastwirth, J. L.(1972) \emph{The Estimation of the Lorenz Curve and Gini Index}, The Review of Economics and Statistics, Vol. 54, No. 3. (August 1972), pp. 306-316.\cr
}
\keyword{datasets}
