\name{michigan}
\alias{michigan}
\docType{data}
\title{Dioxin Levels for Counties in the Upper Peninsula of Michigan}
\description{
 Data contains 16 observations of the Dioxin Levels
 for counties in the Upper Peninsula of Michigan
}
\format{A univariate data set with 16 observations}
\usage{michigan}

\source{The Environmental Protection Agency (EPA) of the State of Michigan}

\keyword{datasets}