\name{popdata}
\alias{popdata}
\docType{data}
\title{Population data in 33 districts 
of the Tennessee Legislature in 1900, 1960 and 1972}
\description{
The Baker v. Carr Case: one-person-one-vote decision. 
Measure of Relative Inequality of Population data in 33 districts 
of the Tennessee Legislature in 1900, 1960 and 1972 (Gastwirth, 1988).
}
\usage{popdata}
\format{
  A data frame with 33 observations on the following 3 variables.
  \describe{
    \item{\code{pop1900}}{population data in 1900}
    \item{\code{pop1960}}{population data in 1960}
    \item{\code{pop1972}}{population data in 1972}
  }
}
\source{
Gastwirth, J. L.(1988) \emph{Statistical Reasoning in Law and
Public Policy Vol 1}, Boston; Toronto, Academic Press.
}
\references{
Gastwirth, J. L.(1988) \emph{Statistical Reasoning in Law and
Public Policy Vol 1}, Boston; Toronto, Academic Press.
}
\keyword{datasets}
