\name{pot}
\alias{pot}
\docType{data}
\title{Apertures of chupa-pots from three Philippine communities }
\description{
The apertures of the chupa pots from the three Philippine locations: 
Dalupa (ApDl), Dangtalan (ApDg) and Paradijon (ApP).
}
\usage{pot}
\format{
A multivariate data set with 343 observations on 2 variables: apertures and 
locations.
}
\details{
  Archaeologists are concerned with the effect that increasing economic activity 
  had on older civilizations. Economic growth and its related economic specialization 
  led to the "standardization hypothesis", i.e. increased production of an item would 
  lead to its becoming more uniform.  Kvamme, Stark and Longacre (1996) focused 
  on earthenware, chupa-pots from three Philippine communities that differ in 
  the way they organize ceramic production. In Dangtalan, pottery is primarily made 
  for household use; in Dalupa there is a non-market barter economy where potters 
  exchange their works. In the village of Paradijon, near the Provincial capital, 
  full-time pottery specialists sell their output to shopkeepers for sale to 
  the general public. 
}
\source{ The data are kindly provided by Professor Kvamme.}
\references{
 Kvamme, K.L., Stark, M.T. and Longacre, M.A. (1996). 
Alternative Procedures for Assessing Standardization in Ceramic Assemblages. 
American Antiquity,61, 116-126.
}

\keyword{datasets}
