% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlmtest.R
\name{rlmtest}
\alias{rlmtest}
\title{Robust Lagrange multiplier test for detecting ARCH/GARCH effect}
\usage{
rlmtest(
  y,
  x = NA,
  a.order = 1,
  sig.level = 0.05,
  crit.type = c("asymptotic", "nonparametric"),
  nonp.method = c("bootstrap", "permutation"),
  num.resamp = 1000,
  num.discard = 0
)
}
\arguments{
\item{y}{A vector that contains univariate time series observations. Missing values are
not allowed.}

\item{x}{A matrix that contains control variable time series observations. Missing values are
not allowed.}

\item{a.order}{Order of the autoregressive model which must be a nonnegative integer number. Default value 1.}

\item{sig.level}{Significance level for testing hypothesis of no ARCH effect. Default value is 0.05.}

\item{crit.type}{Method of obtaining critical values: "asymptotic" (default) or "nonparametric" options.}

\item{nonp.method}{Type of resampling if \code{crit.type}="nonparametric": "bootstrap"
(default) or "permutation".}

\item{num.resamp}{Number of resampling replications if \code{crit.type}="nonparametric". Default number is 1,000.}

\item{num.discard}{Number of bootstrap sample discarded. Default value is 0.}
}
\value{
\item{stats}{Test statistic.}
\item{crit.val}{Ressampling based critical value, with a given significance level \code{sig.level}.}
\item{p.values}{\code{p-value} of the Lagrange multiplier test.}
}
\description{
The function performs two resampling techniques to find critical values of the Lagrange multiplier (LM) test,
namely permutation and bootstrap, as well as the asymptotic parametric test.
}
\examples{

Collect daily high price (in USD) of Bitcoin and gold from June 1st, 2016 to May 31st, 2021,
and create a dataset with log-returned high price of Bitcoin and gold, named as "GoldBTC".
Test for ARCH effects with log-return of high price of Bitcoin and gold with conditions below:
H0: No ARCH effect.
Ha: There exist ARCH effects.

# Let's import dataset "GoldBTC":
data(GoldBTC)

# Fix seed for reproducible simulations:
set.seed(2021)

# We want to test ARCH effect in Gold.
rlmtest(GoldBTC$Gold, crit.type = "nonparametric",
        nonp.method = "bootstrap")

## The list of returned results are below:
## $stat
## [1] 65.84728
## $crit.val
## 95\%
## 1.914901
## $p.value
## [1] 0.003
## Since p-value is 0.003, reject null hypothesis at 0.05,
## that is, ARCH effect exists in gold.

# We want to test for ARCH effects in Bitcoin.
rlmtest(GoldBTC$BTC, crit.type = "nonparametric",
        nonp.method = "bootstrap")

## The list of returned results are below:
## $stat
## [1] 8.189278
## $crit.val
## 95\%
## 3.541198
## $p.value
## [1] 0.013
## Since p-value is 0.013, reject null hypothesis at 0.05,
## that is, ARCH effects exist in Bitcoin.

## We want to test ARCH effect in gold and Bitcoin.
rlmtest(GoldBTC$BTC, GoldBTC$Gold, crit.type = "nonparametric",
        nonp.method = "bootstrap")

## The list of returned results are below:
## $stat
## [1] 9.246483
## $crit.val
## 95\%
## 3.478484
## $p.value
## [1] 0.015
## Since p-value is 0.015, reject null hypothesis at 0.05,
## that is, ARCH effects exist in Bitcoin controling gold.

}
\references{
Engle R F. Autoregressive conditional heteroscedasticity with estimates of the variance of United Kingdom inflation[J].
Econometrica: Journal of the econometric society, 1982: 987-1007.

Gel Y R, Chen B. Robust Lagrange multiplier test for detecting ARCH/GARCH effect using permutation and bootstrap[J].
Canadian Journal of Statistics, 2012, 40(3): 405-426.
}
\author{
Meichen Huang, Tian Jiang, Yulia R. Gel
}
\keyword{Lagrange}
\keyword{bootstrap,}
\keyword{multiplier}
\keyword{resampling}
\keyword{series,}
\keyword{test,}
\keyword{time}
