% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\docType{data}
\name{bias}
\alias{bias}
\title{Prediction Errors ("Biases") of Surface Temperature Forecasts}
\format{
A numeric vector of length 96.
}
\source{
The data were kindly provided by the research group of Professor Clifford 
Mass in the Department of Atmospheric Sciences at the University of Washington. 
Detailed information about the Pacific Northwest prediction effort and the 
associated data archive can be found online at 
\url{https://a.atmos.uw.edu/mm5rt/info.html} and
\url{https://atmos.uw.edu/marka/pnw.html}, respectively.
}
\usage{
data(bias)
}
\description{
Prediction errors of 48-hour ahead MM5 forecasts of surface temperature measured 
at 96 locations in the US Pacific Northwest on 3-January-2000. The prediction error, 
or "bias", is the difference between the forecasted and observed surface
temperature. (MM5 is the fifth-generation Pennsylvania State University -- 
National Center for Atmospheric Research Mesoscale Model.)
}
\keyword{data}
