\name{blackhire}
\alias{blackhire}
\docType{data}
\title{Number of persons being hired by eight professions and two races.}
\description{
The dataset for Cochran-Mantel-Haenszel Chi-square Test (Gastwirth, 1984).}
\usage{blackhire}
\format{
  An array with 2 rows by 2 columns by 8 levels.
  }

\references{
Gastwirth, J. L.(1984) \emph{Statistical Methods for Analyzing Claims of Employment Discrimination}, 
Industrial and Labor Relations Review, Vol. 38, No. 1. (October 1984), pp. 75-86.\cr
}

\keyword{datasets}
