\name{blackhire}
\alias{blackhire}
\docType{data}
\title{Hiring data for eight professions and two races}
\description{
Number of black and white candidates (hired or rejected) for eight professions.}
\usage{blackhire}
\format{An array with 2 rows by 2 columns by 8 levels.}

\references{
Gastwirth, J. L. (1984). Statistical methods for analyzing claims of employment discrimination. \emph{Industrial and Labor Relations Review} 38(1): 75--86.
}

\keyword{datasets}
