% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackhire.R
\docType{data}
\name{blackhire}
\alias{blackhire}
\title{Hiring Data for Eight Professions and Two Races}
\format{
An array with 2 rows by 2 columns by 8 levels.
}
\usage{
data(blackhire)
}
\description{
Number of black and white candidates (hired or rejected) for eight professions
\insertCite{Gastwirth_1984}{lawstat}.
}
\references{
\insertAllCited{}
}
\keyword{data}
