% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd.R
\name{cd}
\alias{cd}
\title{Coefficient of Dispersion -- a Measure of Relative Variability}
\usage{
cd(x)
}
\arguments{
\item{x}{a numeric vector of data values.}
}
\value{
The coefficient of dispersion.
}
\description{
Measure of relative inequality (or relative variation) of the data. 
Coefficient of dispersion (CD) is the ratio of the mean absolute deviation from 
the median (MAAD) to the median of the data. \code{NA}s from the data are omitted.
See \insertCite{Gastwirth_1988v1;textual}{lawstat}  and
\insertCite{Bonett_Seier_2006;textual}{lawstat}.
}
\examples{
## The Baker v. Carr Case: one-person-one-vote decision. 
## Measure of Relative Inequality of Population data in 33 districts 
## of the Tennessee Legislature in 1900 and 1972. See 
## popdata (see Gastwirth, 1988).

data(popdata)
cd(popdata[,"pop1900"])
cd(popdata[,"pop1972"])

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{gini.index}}, \code{\link{j.maad}}
}
\author{
Wallace Hui, Yulia R. Gel, Joseph L. Gastwirth, Weiwen Miao
}
\keyword{homogeneity}
\keyword{variability}
