% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1963.R
\docType{data}
\name{data1963}
\alias{data1963}
\title{Population Size and Number of Senators and Representatives in 1963}
\format{
A data frame with 23 observations on the following 3 variables:
\describe{
    \item{\code{pop1963}}{population in 1963;}
    \item{\code{sen1963}}{number of senators in the district in 1963;}
    \item{\code{rep1963}}{number of representatives in the district in 1963.}
}
}
\source{
\insertCite{Gastwirth_1972;textual}{lawstat}.
}
\usage{
data(data1963)
}
\description{
Number of senators and representatives and population size in 23 districts in 
the United States of America in 1963 \insertCite{Gastwirth_1972}{lawstat}.
}
\references{
\insertAllCited{}
}
\keyword{data}
