% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j.maad.R
\name{j.maad}
\alias{j.maad}
\title{MAAD Robust Standard Deviation}
\usage{
j.maad(x)
}
\arguments{
\item{x}{a numeric vector of data values.}
}
\value{
Robust standard deviation.
}
\description{
Compute average absolute deviation from the sample median, 
which is a consistent robust estimate of the population standard deviation 
for normally distribution data \insertCite{Gastwirth_1982}{lawstat}. 
\code{NA}s from the data are omitted.
}
\examples{
## Sample 100 observations from the standard normal distribution
x = rnorm(100)
j.maad(x)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{cd}}, \code{\link{gini.index}}, \code{\link{rqq}}, 
\code{\link{rjb.test}}, \code{\link{sj.test}}
}
\author{
Wallace Hui, Yulia R. Gel, Joseph L. Gastwirth, Weiwen Miao
}
\keyword{homogeneity}
\keyword{robust}
\keyword{variability}
