% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michigan.R
\docType{data}
\name{michigan}
\alias{michigan}
\title{Dioxin Levels for Counties in the Upper Peninsula of Michigan}
\format{
A numeric vector of length 16.
}
\source{
The Environmental Protection Agency (EPA) of the State of Michigan.
}
\usage{
data(michigan)
}
\description{
Data contains 16 observations of dioxin levels for counties in the Upper Peninsula of Michigan.
}
\keyword{data}
