% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zuni.R
\docType{data}
\name{zuni}
\alias{zuni}
\title{The Zuni Data from the Law Case: Zuni Public School v. United States Department of Education}
\format{
A data frame with 89 observations on 3 variables: \code{District}, 
\code{Revenue}, and \code{Mem} (number of students).
}
\source{
\insertCite{Gastwirth_2006;textual}{lawstat}.
}
\usage{
data(zuni)
}
\description{
Number of students and available revenue per student in each school district in New Mexico.
}
\details{
The Zuni data come from a law case "The Zuni Public School District 
No. 89, Gallup-McKinley County Public School District No. 1, 
Petitioners v. United States Department of Education" concerning whether the 
revenue per pupil satisfied the standard for "equal" expenditures per pupil in the 
state. This classification determines whether most of the federal money given to the 
state under the law goes to the state or to the local school districts.
}
\references{
\insertAllCited{}
}
\keyword{data}
